back-end/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pyserial==3.5